리팩토링, provider  패턴 대신 초기화 방법 고민
