+++ app/layout.tsx
@@ -5,7 +5,6 @@ import blogMetaData from '@/metadata/blogMetaData'
 import { TapRouteMain } from '@/components/ui-layout/tap_system/TapRouteMain'
 import UiContextProvider from '@/components/system/UiContextProvider'
 import { Provider } from 'jotai'
-import { SessionProvider } from '@/components/auth/SessionProvider'
 
 export const metadata: Metadata = {
   // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
@@ -55,15 +54,13 @@ export default function RootLayout ({ children }: { children: React.ReactNode })
     <head></head>
     <body className="overflow-hidden">
     <Provider>
-      <SessionProvider>
-        <UiContextProvider>
+      <UiContextProvider>
 
-          <TapRouteMain
-            page={children}
-          >
-          </TapRouteMain>
-        </UiContextProvider>
-      </SessionProvider>
+        <TapRouteMain
+          page={children}
+        >
+        </TapRouteMain>
+      </UiContextProvider>
     </Provider>
     </body>
     </html>
