+++ src/components/memo/folder_navigator/header/NavigatorHeader.tsx
@@ -2,7 +2,7 @@ import NewMemo from '@/components/memo/folder_navigator/header/NewMemoLink'
 import NewFolder from '@/components/memo/folder_navigator/header/NewFolder'
 import { Search } from '@/components/memo/folder_navigator/header/Search'
 import React, { useContext, useState } from 'react'
-import { type FolderInfo } from '@/outbound/api/models'
+import { type FolderInfo } from '@/api/models'
 import ReferenceSystem from '@/components/memo/folder_navigator/header/ReferenceSystem'
 import { usePathname } from 'next/navigation'
 import { ReferenceModeContext } from '@/components/memo/MemoEditContextProvider'
