+++ src/components/memo/memo_editor/MemoEditWebsocket.ts
@@ -2,7 +2,7 @@ import { useContext, useEffect, useState } from 'react'
 import SockJS from 'sockjs-client'
 import { type Client, type CompatClient, type IMessage, Stomp } from '@stomp/stompjs'
 import { useDebouncedCallback } from 'use-debounce'
-import { type ReferenceInfo } from '@/outbound/api/models'
+import { type ReferenceInfo } from '@/api/models'
 import { ReferenceModeContext } from '@/components/memo/MemoEditContextProvider'
 
 const useStompClient = (
