+++ src/components/ui-layout/tap_system/TabBar.tsx
@@ -1,6 +1,7 @@
 import React, { useContext, useEffect, useRef } from 'react'
-import { type Tab, TabItem } from '@/components/ui-layout/tap_system/TabItem'
+import { TabItem } from '@/components/ui-layout/tap_system/TabItem'
 import { TabBarContext } from '@/components/ui-layout/tap_system/TapRouteMain'
+import { type Tab } from '@/system/application/domain/Tab'
 
 export function TabBar ({ onSelectTab, onRemoveTab, onContextMenu }: {
   onSelectTab: (index: number) => void
