--- src/outbound/api/host.ts
@@ -1,1 +1,0 @@
-export const mainUrl = 'http://localhost:7777/api'
