+++ src/system/application/domain/Tab.ts
@@ -1,0 +1,4 @@
+export interface Tab {
+  name: string
+  context: string
+}
