+++ src/system/application/usecase/TabUseCases.ts
@@ -1,25 +1,7 @@
 import type React from 'react'
 import { useCallback, useEffect, useState } from 'react'
-import { type Tab } from '@/components/ui-layout/tap_system/TabItem'
-
-const restoreTabsFromLocalStorage = (path: string): any => {
-  const savedTabs = localStorage.getItem('tabs')
-  const tabsList = (savedTabs != null) ? JSON.parse(savedTabs) : null
-  if (path.startsWith('/login/oauth2')) {
-    if (tabsList == null || tabsList.length === 0) {
-      return [{ name: '/', context: '/' }]
-    } else {
-      return tabsList
-    }
-  }
-  if (tabsList == null || (tabsList.length === 0 && path !== '/empty')) {
-    return path === '/empty' ? [] : [{ name: path, context: path }]
-  }
-  if ((tabsList.length === 0 && path === '/empty') || (tabsList.length !== 0 && path === '/empty')) {
-    return []
-  }
-  return tabsList
-}
+import { restoreTabsFromLocalStorage } from '@/system/infra/localstorage/TabLocalStorage'
+import { type Tab } from '@/system/application/domain/Tab'
 
 export function useTabs (initialPath: string): {
   tabs: Tab[]
