+++ src/utils/host.ts
@@ -1,0 +1,1 @@
+export const LocalHost = 'http://localhost:7777/api'
